+++ Scripts/Pickups/AirDashItem.cs
@@ -1,0 +1,11 @@
+// using Godot;
+// using System;
+
+// public partial class AirDashItem : PickupArea2D
+// {
+// 	protected override void ApplyPickupEffect(Player player)
+// 	{
+// 		player.AirDashEnabled = true;
+// 		GD.Print("Air Dash enabled!");
+// 	}
+// }
