+++ Scripts/Pickups/DoubleJumpItem.cs
@@ -1,0 +1,11 @@
+// using Godot;
+// using System;
+
+// public partial class DoubleJumpItem : PickupArea2D
+// {
+//     protected override void ApplyPickupEffect(Player player)
+//     {
+//         player.DoubleJumpEnabled = true;
+//         GD.Print("Double Jump enabled!");
+//     }
+// }
